requirements.txt
amqp==5.3.1
anyascii==0.3.3
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==5.2.9
django-filter==25.2
django-jazzmin==3.0.1
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-rest-framework==0.1.0
django-stubs-ext==5.2.8
django-taggit==6.1.0
django-tasks==0.9.0
django-treebeard==4.8.0
djangorestframework==3.16.1
draftjs_exporter==5.1.0
drf-spectacular==0.29.0
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filetype==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
laces==0.1.2
MarkupSafe==3.0.3
modelsearch==1.1.1
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
pillow_heif==1.1.1
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-telegram-bot==20.7
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlparse==0.5.4
telepath==0.3.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.1
vine==5.1.0
wagtail==7.2.1
wcwidth==0.2.14
Willow==1.12.0
uvicorn
python-dotenv
